Update README.md
In this updated document, you will find an overview of the project, including the process implemented to achieve the business objectives. Uncover the valuable insights obtained from the data, along with the methodologies employed to address these insights. Explore the challenges encountered throughout the project and the quality assurance (QA) process employed to ensure the accuracy and validity of our findings. Lastly, discover future considerations that can be pursued. To gain a more comprehensive understanding, I invite you to explore the README for detailed information and in-depth insights.